chore(test_indev): merge indev testcase
